Add basic Section element
Often when laying out the contents of a web page, the page is broken
up into sections, this is where the Section element comes in handy. The
element accepts all content & container elements as a child. The markup
uses the `section` tag with a child DOM element that houses all the
children.

Below is an example of the default markup that's rendered to
the page:

```
<section class="section">
	<div class="section__content">
		...
	</div>
</section>
```